All components with the corresponding CSS work
What all works?
Login, Events(Register part pending as DB not implemented for it)
Internships display properly(DB not implemented)

What was done to segregate CSS
-> The container class of login page conflicted with the container of
other components so created a module for the components of the login
page. But don't know somehow the container of NavBar started
malfunctioning so I created a separate eventcard module and individually
applied container CSS to necessary components. Also there was a star *
property for the login page which got applied despite that being a
module. So tried fitting its properties in a new div named LOGINPAGE.