Make login work using token
Some changes made:
1. Added role column in database the query, for which is in queries.sql

What is happening?
login.js in server will send a token on successful login. The role will
be obtained from the new role column created. The token sent will have
this encrypted role in it, which will then be decoded and the user will
be directed accordingly.